requirements.txt
absl-py==1.4.0
attrs==19.3.0
Automat==0.8.0
blessings==1.6
blinker==1.4
braceexpand==0.1.7
cachetools==5.3.1
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cloud-init==23.1.2
cmake==3.27.2
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
entrypoints==0.3
filelock==3.12.2
fsspec==2023.6.0
ftfy==6.1.1
google-auth==2.22.0
google-auth-oauthlib==1.0.0
gpustat==0.6.0
grpcio==1.57.0
httplib2==0.14.0
huggingface-hub==0.16.4
hyperlink==19.0.0
idna==2.8
importlib-metadata==6.8.0
incremental==16.10.1
Jinja2==2.10.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lit==16.0.6
macaroonbakery==1.3.1
Markdown==3.4.4
MarkupSafe==2.1.3
more-itertools==4.2.0
mpmath==1.3.0
netifaces==0.10.4
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-ml-py3==7.352.0
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.1.0
open-clip-torch==2.20.0
packaging==23.1
pandas==2.0.3
pexpect==4.6.0
Pillow==10.0.0
protobuf==4.24.1
psutil==5.5.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycups==1.9.73
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyRFC3339==1.1
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.8
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
pytz==2023.3
PyYAML==5.3.1
regex==2023.8.8
requests==2.22.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
rsa==4.9
safetensors==0.3.2
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
sentencepiece==0.1.99
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sos==4.3
ssh-import-id==5.10
sympy==1.12
systemd-python==234
tensorboard==2.14.0
tensorboard-data-server==0.7.1
timm==0.9.5
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
triton==2.0.0
Twisted==18.9.0
typing-extensions==4.7.1
tzdata==2023.3
ubuntu-advantage-tools==27.10
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
wadllib==1.3.3
wcwidth==0.2.6
webdataset==0.2.48
werkzeug==2.3.7
xkit==0.0.0
zipp==1.0.0
zope.interface==4.7.1
